Vision.Image

Description

Images are manipulated by their Image and MaskedImage type-class instances.

The Manifest representation uses an internal Vector to represent the image whereas the Delayed representation uses a function to generate pixels. Most transformation functions are generic to both representations in the way they apply to any type which implements the type-classes.

The Delayed image should be used as intermediate representations of complex image transformations.

Please refer to our README file for a detailed usage and examples.

Documentation

module Vision.Image.Class

module Vision.Image.Conversion

module Vision.Image.Grey

module Vision.Image.Filter

module Vision.Image.HSV

module Vision.Image.Interpolate

module Vision.Image.Mutable

module Vision.Image.Parallel

module Vision.Image.RGB

module Vision.Image.RGBA

module Vision.Image.Threshold

module Vision.Image.Transform

module Vision.Image.Type