Vision.Image.HSV.Type

Documentation

type HSV = Manifest HSVPixel

24 bits (3 * 8 bits) HSV image.

The Hue value is in [0..179], Saturation in [0..255] and Value in [0..255].

This image type is more respectful to human eye perception of colors and can be converted (using convert) from RGB images.

Uses http://en.wikipedia.org/wiki/HSL_and_HSV equations to convert from and to RGB.

data HSVPixel

Constructors

HSVPixel

Fields

hsvHue :: !Word8

hsvSat :: !Word8

hsvValue :: !Word8

type HSVDelayed = Delayed HSVPixel