Vision.Image.Parallel

Documentation

computeP :: (Monad m, Image i, Storable (ImagePixel i)) => i -> m (Manifest (ImagePixel i))

Parallel version of compute.

Computes the value of an image into a manifest representation in parallel.

The monad ensures that the image is fully evaluated before continuing.