Vision.Image.Type

Manifest images

data Manifest p

Stores the image content in a Vector.

Constructors

Manifest

Fields

manifestSize :: !Size

manifestVector :: !(Vector p)

Delayed images

data Delayed p

A delayed image is an image which is constructed using a function.

Usually, a delayed image maps each of its pixels over another image. Delayed images are useful by avoiding intermediate images in a transformation pipeline of images or by avoiding the computation of the whole resulting image when only a portion of its pixels will be accessed.

Constructors

Delayed

Fields

delayedSize :: !Size

delayedFun :: !(Point -> p)

Delayed masked images

data DelayedMask p

Constructors

DelayedMask

Fields

delayedMaskSize :: !Size

delayedMaskFun :: !(Point -> Maybe p)

Conversion and type helpers

delay :: Image i => i -> Delayed (ImagePixel i)

Delays an image in its delayed representation.

compute :: (Image i, Storable (ImagePixel i)) => i -> Manifest (ImagePixel i)

Computes the value of an image into a manifest representation.

delayed :: Delayed p -> Delayed p

Forces an image to be in its delayed represenation. Does nothing.

manifest :: Manifest p -> Manifest p

Forces an image to be in its manifest represenation. Does nothing.