Vision.Primitive

Documentation

module Vision.Primitive.Shape

type Point = DIM2

Coordinates inside the image.

Can be constructed using ix2. The first parameter is the y coordinate while the second is the x coordinate (i.e. ix2 y x). Image origin (ix2 0 0) is located in the upper left corner.

type Size = DIM2

Size of an object.

Can be constructed using ix2. The first parameter is the height while the second is the width (i.e. ix2 h w).

data Rect

Constructors

Rect

Fields

rX :: !Int

rY :: !Int

rWidth :: !Int

rHeight :: !Int

data RPoint

Rational coordinates used for interpolations.

Constructors

RPoint

Fields

rpX :: !RatioInt

rpY :: !RatioInt