Vision.Primitive.Shape

Description

Shapes are similar to what you could found in repa. Shape are used both for indexes and shapes.

To create a shape/index, use the ix1, ix2, ix3 ... helpers :

size = ix2 200 100

To pull values from a shape, use the Z and :. constructors :

Z :. h :. w = size

Documentation

class Eq sh => Shape sh where

Class of types that can be used as array shapes and indices.

Methods

shapeRank :: sh -> Int

Gets the number of dimensions in a shape.

shapeLength :: sh -> Int

Gets the total number of elements in an array of this shape.

shapeZero :: sh

Gives the first index of an array.

shapeSucc

Arguments

:: sh
Shape of the array.

-> sh
Index.

-> sh

Gives the successor of an index, given the shape of the array.

toLinearIndex

Arguments

:: sh
Shape of the array.

-> sh
Index into the array.

-> Int

Convert an index into its equivalent flat, linear, row-major version.

fromLinearIndex

Arguments

:: sh
Shape of the array.

-> Int
Index into linear representation.

-> sh

Inverse of toLinearIndex.

shapeList :: sh -> [sh]

Return the ascending list of indexes for the given shape.

inShape

Arguments

:: sh
Shape of the array.

-> sh
Index to check for.

-> Bool

Check whether an index is within a given shape.

data Z

An index of dimension zero.

Constructors

Z

data tail :. head infixl 3

Our index type, used for both shapes and indices.

Constructors

!tail :. !head infixl 3

Common dimensions.

type DIM0 = Z

type DIM1 = DIM0 :. Int

type DIM2 = DIM1 :. Int

type DIM3 = DIM2 :. Int

type DIM4 = DIM3 :. Int

type DIM5 = DIM4 :. Int

type DIM6 = DIM5 :. Int

type DIM7 = DIM6 :. Int

type DIM8 = DIM7 :. Int

type DIM9 = DIM8 :. Int

Helpers

ix1 :: Int -> DIM1

Helper for index construction.

Use this instead of explicit constructors like (Z :. (x :: Int)) The this is sometimes needed to ensure that x is constrained to be in Int.

ix2 :: Int -> Int -> DIM2

ix3 :: Int -> Int -> Int -> DIM3

ix4 :: Int -> Int -> Int -> Int -> DIM4

ix5 :: Int -> Int -> Int -> Int -> Int -> DIM5

ix6 :: Int -> Int -> Int -> Int -> Int -> Int -> DIM6

ix7 :: Int -> Int -> Int -> Int -> Int -> Int -> Int -> DIM7

ix8 :: Int -> Int -> Int -> Int -> Int -> Int -> Int -> Int -> DIM8

ix9 :: Int -> Int -> Int -> Int -> Int -> Int -> Int -> Int -> Int -> DIM9