Vision.Image.Conversion

Documentation

class Convertible a b where

A typeclass that represents something that can be converted. A Convertible a b instance represents an a that can be converted to a b.

Methods

safeConvert :: a -> ConvertResult b

Convert a to b, returning Right on success and Left on error. For a simpler interface, see convert.

convert :: Convertible a b => a -> b

Convert from one type of data to another. Raises an exception if there is an error with the conversion. For a function that does not raise an exception in that case, see safeConvert.